Fixes #5775: Implement issue #5775
